Added can to control system(still need to test but i need to save the code)

+++ cruise_control/src/control/pid.h
@@ -34,12 +34,5 @@ struct pid
 };
 
 
-
-
-
-
-
-
-
-void set_pid_target_rpm(uint32_t target);
+void set_pid_target_rpm(uint16_t target);
 void start_pid_controller();
